requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
dateparser==1.2.0
demjson3==3.0.6
exceptiongroup==1.2.0
facebook-scraper @ git+https://github.com/moda20/facebook-scraper.git@b636187d38b6a451e7165a7324571ee411ff65ce
fake-useragent==1.5.1
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.1
iniconfig==2.0.0
lxml==5.1.0
mysql-connector-python==8.3.0
packaging==24.0
parse==1.20.1
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.28.0
w3lib==2.1.2
websockets==10.4
zipp==3.18.1
